# docs/openapi/policy-enforcer-openapi.yaml
openapi: 3.0.3
info:
  title: DYNAMOS Policy Enforcer API
  description: |
    API for policy enforcement in the DYNAMOS system.
    
    This API provides two levels of abstraction:
    
    ## Policy Enforcer API (`/policy-enforcer/*`)
    High-level, reasoner-agnostic interface for policy validation. This API is designed
    to work with different reasoning engines (eFLINT, Symboleo, JSON-based, etc.) through
    a common interface. Use these endpoints for:
    - **Allowed Clauses**: Query all allowed request types, datasets, archetypes, and compute
      providers for a specific requester at an organization in a single call
    - **Request Validation**: Check if a specific data request is permitted by the policy
    
    ## eFLINT API (`/eflint/*`)
    Low-level interface for direct eFLINT server management. Use these endpoints for:
    - **Instance Management**: Start, stop, and monitor eFLINT server instances
    - **Commands**: Send raw commands to the eFLINT server
    - **State Management (POC)**: Export, import, and checkpoint eFLINT execution graphs
    
    ## Architecture
    ```
    ┌─────────────────────────────────────────────────────────────────────────┐
    │                           HTTP/gRPC Layer                                │
    ├─────────────────────────────────────────────────────────────────────────┤
    │  Policy Enforcer API              │  eFLINT API                         │
...
                     │                                       │
...
servers:
...
paths:
...
  /api/v1/eflint/state:
    get:
      summary: Get current state
      description: |
        Retrieves the current execution graph state of the eFLINT instance.
        If instance_id is provided, returns the state of that specific pool instance.
      operationId: getState
      tags:
        - State Management
      parameters:
        - $ref: '#/components/parameters/InstanceIdParam'
      responses:
        '200':
          description: State retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StateResponse'
        '503':
          description: Instance is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/export:
    post:
      summary: Export state
      description: |
        Exports the current eFLINT execution graph state for later import.
        Returns a SavedState object with metadata that can be stored and restored.
      operationId: exportState
      tags:
        - State Management
      responses:
        '200':
          description: State exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStateResponse'
        '503':
          description: Instance is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/import:
    post:
      summary: Import state
      description: |
        Imports a previously exported eFLINT execution graph state.
        
        > **Note**: Due to limitations in the eFLINT server's load-export functionality,
        > full state restoration may not work in all cases.
      operationId: importState
      tags:
        - State Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportStateRequest'
      responses:
        '200':
          description: State imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request - state is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Instance is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/checkpoint:
    post:
      summary: Create checkpoint
      description: Creates a named checkpoint of the current state that can be restored later
      operationId: createCheckpoint
      tags:
        - State Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckpointRequest'
            example:
              name: "before-test"
      responses:
        '200':
          description: Checkpoint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointCreatedResponse'
        '400':
          description: Bad request - name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Instance is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/checkpoint/restore:
    post:
      summary: Restore checkpoint
      description: |
        Restores a previously created checkpoint.
        
        > **Note**: Due to a bug in the eFLINT server, full state restoration may not work.
        > In that case, the instance will be restarted to the initial model state.
      operationId: restoreCheckpoint
      tags:
        - State Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckpointRequest'
            example:
              name: "before-test"
      responses:
        '200':
          description: Checkpoint restored successfully (or partially with warning)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointRestoredResponse'
        '400':
          description: Bad request - name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Instance is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/checkpoints:
    get:
      summary: List checkpoints
      description: Lists all available checkpoints
      operationId: listCheckpoints
      tags:
        - State Management
      responses:
        '200':
          description: Checkpoints listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/eflint/state/checkpoint/{name}:
    delete:
      summary: Delete checkpoint
      description: Deletes a checkpoint by name
      operationId: deleteCheckpoint
      tags:
        - State Management
      parameters:
        - name: name
          in: path
          required: true
          description: The name of the checkpoint to delete
          schema:
            type: string
          example: "before-test"
      responses:
        '200':
          description: Checkpoint deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointDeletedResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # Health Endpoints
  # ---------------------------------------------------------------------------
...
components:
...
  parameters:
...
    InstanceIdParam:
      name: instance_id
      in: query
      required: false
      description: Optional pool instance ID to target. If omitted, the default instance is used.
      schema:
        type: string
      example: "instance-0"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    StateResponse:
      type: object
      properties:
        state:
          type: object
          description: The current execution graph state
    SavedState:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the saved state
          example: "state-1704470400000000000"
        model_location:
          type: string
          description: The model location when the state was saved
          example: "/eflint/empty.eflint"
        graph:
          type: object
          description: The eFLINT execution graph
        saved_at:
          type: string
          format: date-time
          description: When the state was saved
          example: "2026-01-05T15:30:00Z"
    ExportStateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        state:
          $ref: '#/components/schemas/SavedState'
    ImportStateRequest:
      type: object
      required:
        - state
      properties:
        state:
          $ref: '#/components/schemas/SavedState'
    CheckpointRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The name of the checkpoint
          example: "before-test"
    CheckpointCreatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        checkpoint:
          type: string
          description: The name of the created checkpoint
          example: "before-test"
        state_id:
          type: string
          description: The ID of the saved state
          example: "state-1704470400000000000"
        saved_at:
          type: string
          format: date-time
          example: "2026-01-05T15:30:00Z"
    CheckpointRestoredResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "checkpoint restored successfully"
        restored:
          type: string
          description: The name of the restored checkpoint
          example: "before-test"
    CheckpointListResponse:
      type: object
      properties:
        checkpoints:
          type: array
          items:
            type: string
          description: List of checkpoint names
          example: ["before-test", "initial-state"]
    CheckpointDeletedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "checkpoint deleted successfully"
        deleted:
          type: string
          description: The name of the deleted checkpoint
          example: "before-test"

    # -------------------------------------------------------------------------
    # Common Schemas
    # -------------------------------------------------------------------------
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "no instance running"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "operation completed successfully"

# -----------------------------------------------------------------------------
# Tags
# -----------------------------------------------------------------------------
...